+++ App.tsx
@@ -4,9 +4,8 @@ import { PersistGate } from 'redux-persist/integration/react';
 
 import SettingProvider from 'contexts/SettingProvider';
 import 'packages/localization';
-import AppBar from 'packages/components/AppBar';
-import Drawer, { globalDrawerRef } from 'packages/components/Drawer';
-import GestureRecognizer from 'packages/components/Drawer/GestureRecognizer';
+import Drawer, { globalDrawerRef } from 'packages/uikit/components/Drawer';
+import GestureRecognizer from 'packages/uikit/components/Drawer/GestureRecognizer';
 import { BaseProvider, IBaseConfig } from 'packages/core';
 import { onSwipeRight } from 'packages/utils/gestureHandle';
 import RootStacks from 'stacks';
@@ -21,13 +20,13 @@ function App() {
     return (
         <BaseProvider config={config}>
             <SettingProvider>
-                <GestureRecognizer onSwipeRight={(state) => onSwipeRight(state)} style={{ flex: 1 }} >
-                <Provider store={store}>
-                    <Drawer showAvatar ref={globalDrawerRef} />
-                    <PersistGate loading={null} persistor={persistor}>
-                        <RootStacks />
-                    </PersistGate>
-                </Provider>
+                <GestureRecognizer onSwipeRight={(state) => onSwipeRight(state)} style={{ flex: 1 }}>
+                    <Provider store={store}>
+                        <Drawer showAvatar ref={globalDrawerRef} />
+                        <PersistGate loading={null} persistor={persistor}>
+                            <RootStacks />
+                        </PersistGate>
+                    </Provider>
                 </GestureRecognizer>
             </SettingProvider>
         </BaseProvider>
