--- src/packages/components/Modal/index.ts
@@ -1,15 +1,0 @@
-import { BaseModal } from './modal';
-
-export type { OnSwipeCompleteParams } from './modal';
-export type {
-  AnimationEvent,
-  Animations,
-  SupportedAnimation,
-  Orientation,
-  Direction,
-  PresentationStyle,
-  OnOrientationChange,
-  GestureResponderEvent,
-} from './types';
-
-export default BaseModal;
