+++ src/packages/utils/gestureHandle.ts
@@ -1,10 +1,8 @@
-import { globalDrawer } from 'packages/components/Drawer';
+import { globalDrawer } from 'packages/uikit/components/Drawer';
 
 const onSwipeRight = (gestureState) => {
     console.log(gestureState);
-    globalDrawer.open()
-}
+    globalDrawer.open();
+};
 
-export {
-    onSwipeRight,
-}
+export { onSwipeRight };
